move terrain component to useTerrain hook
