--- src/components/SphereMap/Terrain.tsx
@@ -1,49 +1,0 @@
-import { useMap } from "react-map-gl/maplibre"
-import { useEffect } from "react"
-import { useAppSelector } from "@/store/hooks"
-import { selectExaggeration } from "@/store/terrain"
-
-export type TerrainProps = {
-    mapId: string
-}
-
-export const Terrain: React.FC<TerrainProps> = ({ mapId }) => {
-    const { [mapId]: ref } = useMap()
-    const exaggeration = useAppSelector(selectExaggeration)
-
-    useEffect(() => {
-        const map = ref?.getMap()
-        if (!map) {
-            return
-        }
-
-        const setupTerrain = () => {
-            // add the DEM source as a terrain layer with exaggerated height
-            map.setTerrain({
-                source: "mapbox-dem",
-                exaggeration,
-            })
-        }
-
-        if (map.isStyleLoaded()) {
-            setupTerrain()
-
-            return () => {
-                if (map.isStyleLoaded()) {
-                    map.setTerrain(null)
-                }
-            }
-        }
-
-        map.on("load", setupTerrain)
-
-        return () => {
-            map.off("load", setupTerrain)
-            if (map.isStyleLoaded()) {
-                map.setTerrain(null)
-            }
-        }
-    }, [ref, exaggeration])
-
-    return null
-}
